keep reporting capabilities in impactutils
