+++ utilities/ImpactUtilities.groovy
@@ -13,7 +13,6 @@ import java.util.regex.*
 @Field BuildProperties props = BuildProperties.getInstance()
 @Field def gitUtils= loadScript(new File("GitUtilities.groovy"))
 @Field def buildUtils= loadScript(new File("BuildUtilities.groovy"))
-@Field def reportingUtils= loadScript(new File("ReportingUtilities.groovy"))
 @Field String hashPrefix = ':githash:'
 @Field def resolverUtils
 
@@ -49,7 +48,7 @@ def createImpactBuildList(RepositoryClient repositoryClient) {
 	// Perform analysis and build report of external impacts
 	if (props.reportExternalImpacts && props.reportExternalImpacts.toBoolean()){
 		if (props.verbose) println "*** Perform analysis and report external impacted files for changed files."
-		reportingUtils.reportExternalImpacts(repositoryClient, changedFiles)
+		reportExternalImpacts(repositoryClient, changedFiles)
 	}
 
 
@@ -221,7 +220,7 @@ def createMergeBuildList(RepositoryClient repositoryClient){
 	// Perform analysis and build report of external impacts
 	if (props.reportExternalImpacts && props.reportExternalImpacts.toBoolean()){
 		if (props.verbose) println "*** Analyze and report external impacted files."
-		reportingUtils.reportExternalImpacts(repositoryClient, changedFiles)
+		reportExternalImpacts(repositoryClient, changedFiles)
 	}
 	
 	return [buildSet, deletedFiles]
@@ -633,6 +632,109 @@ def generateConcurrentChangesReports(Set<String> buildList, Set<String> concurre
 			}
 		}
 	}
+}
+
+/**
+ * Method to query the DBB collections with a list of changed files
+ * Configured through reportExternalImpacts* build properties
+ */
+
+def reportExternalImpacts(RepositoryClient repositoryClient, Set<String> changedFiles){
+	// query external collections to produce externalImpactList
+
+	Map<String,HashSet> collectionImpactsSetMap = new HashMap<String,HashSet>() // <collection><List impactRecords>
+	Set<String> impactedFiles = new HashSet<String>()
+
+	if (props.reportExternalImpactsAnalysisDepths == "simple" || props.reportExternalImpactsAnalysisDepths == "deep"){
+
+		// calculate and collect external impacts
+		changedFiles.each{ changedFile ->
+
+			List<PathMatcher> fileMatchers = createPathMatcherPattern(props.reportExternalImpactsAnalysisFileFilter)
+
+			// check that file is on reportExternalImpactsAnalysisFileFilter
+			if(matches(changedFile, fileMatchers)){
+
+				// get directly impacted candidates first
+				(collectionImpactsSetMap, impactedFiles) = calculateLogicalImpactedFiles(changedFile, collectionImpactsSetMap, repositoryClient)
+
+				// get impacted files of idenfied impacted files
+				if (props.reportExternalImpactsAnalysisDepths == "deep") {
+					impactedFiles.each{ impactedFile ->
+						(collectionImpactsSetMap, impactedFiles) = calculateLogicalImpactedFiles(impactedFile, collectionImpactsSetMap, repositoryClient)
+					}
+				}
+
+				// generate reports by collection / application
+				collectionImpactsSetMap.each{ entry ->
+					externalImpactList = entry.value
+					if (externalImpactList.size()!=0){
+						// write impactedFiles per application to build workspace
+						String impactListFileLoc = "${props.buildOutDir}/externalImpacts_${entry.key}.${props.buildListFileExt}"
+						if (props.verbose) println("*** Writing report of external impacts to file $impactListFileLoc")
+						File impactListFile = new File(impactListFileLoc)
+						String enc = props.logEncoding ?: 'IBM-1047'
+						impactListFile.withWriter(enc) { writer ->
+							externalImpactList.each { file ->
+								// if (props.verbose) println file
+								writer.write("$file\n")
+							}
+						}
+					}
+				}
+
+			} else {
+				if (props.verbose) println("*** Analysis and reporting has been skipped for changed file $changedFile due to build framework configuration (see configuration of build property reportExternalImpactsAnalysisFileFilter)")
+			}
+		}
+	}
+	else {
+		println("*! build property reportExternalImpactsAnalysisDepths has an invalid value : ${props.reportExternalImpactsAnaylsisDepths} , valid: simple | deep")
+	}
+}
+
+/*
+ * Used to  
+ */
+
+def calculateLogicalImpactedFiles(String changedFile, Map<String,HashSet> collectionImpactsSetMap, RepositoryClient repositoryClient) {
+
+	// local matchers to inspect files and collections
+	List<Pattern> collectionMatcherPatterns = createMatcherPatterns(props.reportExternalImpactsCollectionPatterns)
+
+	// will be returned
+	Set<String> impactedFiles = new HashSet<String>()
+
+	String memberName = CopyToPDS.createMemberName(changedFile)
+
+	def ldepFile = new LogicalDependency(memberName, null, null);
+	repositoryClient.getAllCollections().each{ collection ->
+		String cName = collection.getName()
+		if(matchesPattern(cName,collectionMatcherPatterns)){ // find matching collection names
+			def Set<String> externalImpactList = collectionImpactsSetMap.get(cName) ?: new HashSet<String>()
+			def logicalImpactedFiles = repositoryClient.getAllLogicalFiles(cName, ldepFile);
+
+			logicalImpactedFiles.each{ logicalFile ->
+				if (props.verbose) println("*** Changed file $changedFile has a potential external impact on logical file ${logicalFile.getLname()} (${logicalFile.getFile()}) in collection ${cName} ")
+				def impactRecord = "${logicalFile.getLname()} \t ${logicalFile.getFile()} \t ${cName}"
+
+				if (cName != props.applicationCollectionName && cName != props.applicationOutputsCollectionName){ // we can exclude internal impacted files
+					externalImpactList.add(impactRecord)
+				}
+				impactedFiles.add(logicalFile.getFile())
+			}
+			// adding updated record
+			collectionImpactsSetMap.put(cName, externalImpactList)
+		}
+		else{
+			//if (props.verbose) println("$cName does not match pattern: $collectionMatcherPatterns")
+		}
+	}
+
+	return [
+		collectionImpactsSetMap,
+		impactedFiles
+	]
 }
 
 def createImpactResolver(String changedFile, String rules, RepositoryClient repositoryClient) {
